+++ src/WhatToCook.WebApp/Program.cs
@@ -1,4 +1,5 @@
 using WhatToCook.Application.Infrastructure;
+using WhatToCook.Application.Infrastructure.Seeds;
 using WhatToCook.Application.Services;
 
 var builder = WebApplication.CreateBuilder(args);
@@ -10,16 +11,20 @@ builder.Services.AddSwaggerGen();
 builder.Services.RegisterApplicationServices();
 var app = builder.Build();
 
+if (app.Environment.IsDevelopment())
+{
+	app.UseSwagger();
+	app.UseSwaggerUI();
+	app.Services.SeedRecipes();
+}
 // Configure the HTTP request pipeline.
 if (!app.Environment.IsDevelopment())
 {
-    // The default HSTS value is 30 days. You may want to change this for production scenarios, see https://aka.ms/aspnetcore-hsts.
-    app.UseHsts();
+	// The default HSTS value is 30 days. You may want to change this for production scenarios, see https://aka.ms/aspnetcore-hsts.
+	app.UseHsts();
 }
 else
 {
-    app.UseSwagger();
-    app.UseSwaggerUI();
 }
 app.UseStaticFiles();
 app.UseHttpsRedirection();
@@ -27,8 +32,8 @@ app.UseStaticFiles();
 app.UseRouting();
 
 app.MapControllerRoute(
-    name: "default",
-    pattern: "{controller}/{action=Index}/{id?}");
+	name: "default",
+	pattern: "{controller}/{action=Index}/{id?}");
 
 app.MapFallbackToFile("index.html");
